# contracts/openapi/konsider-api-3.0.json
{
  "components": {
    "schemas": {
...
      "CatalogV3Response": {
        "additionalProperties": false,
        "properties": {
          "api_contract_version": {
            "const": "konsider-api-3.0",
            "default": "konsider-api-3.0",
            "title": "Api Contract Version",
            "type": "string"
          },
          "catalog_schema_version": {
            "title": "Catalog Schema Version",
            "type": "string"
          },
          "countries": {
            "items": {
              "$ref": "#/components/schemas/GeographicEntityResponse"
            },
            "title": "Countries",
            "type": "array"
          },
          "country_coverage": {
            "items": {
              "$ref": "#/components/schemas/CountryCoverageResponse"
            },
            "title": "Country Coverage",
            "type": "array"
          },
          "coverage_policy_version": {
            "title": "Coverage Policy Version",
            "type": "string"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CatalogCriterionV3Response"
            },
            "title": "Criteria",
            "type": "array"
          },
          "preference_presets": {
            "items": {
              "$ref": "#/components/schemas/PreferencePresetResponse"
            },
            "title": "Preference Presets",
            "type": "array"
          },
          "release_id": {
            "title": "Release Id",
            "type": "string"
          },
          "release_schema_version": {
            "title": "Release Schema Version",
            "type": "string"
          },
          "scoring_method_versions": {
            "items": {
              "type": "string"
            },
            "title": "Scoring Method Versions",
            "type": "array"
          },
          "stable_universe_id": {
            "title": "Stable Universe Id",
            "type": "string"
          }
        },
        "required": [
          "release_id",
          "release_schema_version",
          "catalog_schema_version",
          "scoring_method_versions",
          "coverage_policy_version",
          "stable_universe_id",
          "countries",
          "country_coverage",
          "criteria",
          "preference_presets"
        ],
        "title": "CatalogV3Response",
        "type": "object"
      },
...
      "ComparisonV3Response": {
        "additionalProperties": false,
        "properties": {
          "api_contract_version": {
            "const": "konsider-api-3.0",
            "default": "konsider-api-3.0",
            "title": "Api Contract Version",
            "type": "string"
          },
          "assessments": {
            "$ref": "#/components/schemas/AssessmentsV3Response"
          },
          "catalog_schema_version": {
            "title": "Catalog Schema Version",
            "type": "string"
          },
          "countries": {
            "items": {
              "$ref": "#/components/schemas/ComparedCountryV3Response"
            },
            "title": "Countries",
            "type": "array"
          },
          "criterion_rows": {
            "items": {
              "$ref": "#/components/schemas/ComparisonCriterionRowV3Response"
            },
            "title": "Criterion Rows",
            "type": "array"
          },
          "normalized_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Normalized Weights",
            "type": "object"
          },
          "release_id": {
            "title": "Release Id",
            "type": "string"
          },
          "release_schema_version": {
            "title": "Release Schema Version",
            "type": "string"
          },
          "requested_country_entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Requested Country Entity Ids",
            "type": "array"
          },
          "resolved_preference_preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Preference Preset Id"
          },
          "scoring_method_versions": {
            "items": {
              "type": "string"
            },
            "title": "Scoring Method Versions",
            "type": "array"
          }
        },
        "required": [
          "release_id",
          "release_schema_version",
          "catalog_schema_version",
          "scoring_method_versions",
          "resolved_preference_preset_id",
          "normalized_weights",
          "assessments",
          "requested_country_entity_ids",
          "countries",
          "criterion_rows"
        ],
        "title": "ComparisonV3Response",
        "type": "object"
      },
...
      "CountryDetailsV3Response": {
        "additionalProperties": false,
        "properties": {
          "api_contract_version": {
            "const": "konsider-api-3.0",
            "default": "konsider-api-3.0",
            "title": "Api Contract Version",
            "type": "string"
          },
          "assessments": {
            "$ref": "#/components/schemas/AssessmentsV3Response"
          },
          "catalog_schema_version": {
            "title": "Catalog Schema Version",
            "type": "string"
          },
          "country": {
            "$ref": "#/components/schemas/GeographicEntityResponse"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CountryCriterionDetailV3Response"
            },
            "title": "Criteria",
            "type": "array"
          },
          "feasibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TfcCountryAssessmentV2Response"
              },
              {
                "type": "null"
              }
            ]
          },
          "normalized_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Normalized Weights",
            "type": "object"
          },
          "opportunity_filters": {
            "items": {
              "$ref": "#/components/schemas/OpportunityFilterEvidenceSummaryResponse"
            },
            "title": "Opportunity Filters",
            "type": "array"
          },
          "release_id": {
            "title": "Release Id",
            "type": "string"
          },
          "release_schema_version": {
            "title": "Release Schema Version",
            "type": "string"
          },
          "resolved_preference_preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Preference Preset Id"
          },
          "scoring_method_versions": {
            "items": {
              "type": "string"
            },
            "title": "Scoring Method Versions",
            "type": "array"
          }
        },
        "required": [
          "release_id",
          "release_schema_version",
          "catalog_schema_version",
          "scoring_method_versions",
          "resolved_preference_preset_id",
          "normalized_weights",
          "assessments",
          "country",
          "criteria",
          "opportunity_filters"
        ],
        "title": "CountryDetailsV3Response",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GeographicEntityResponse": {
        "additionalProperties": false,
        "properties": {
          "country_codes": {
            "items": {
              "type": "string"
            },
            "title": "Country Codes",
            "type": "array"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "enum": [
              "COUNTRY",
              "CITY",
              "METRO",
              "REGION",
              "SERVICE_AREA"
            ],
            "title": "Entity Type",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          }
        },
        "required": [
          "entity_id",
          "entity_type",
          "display_name",
          "country_codes"
        ],
        "title": "GeographicEntityResponse",
        "type": "object"
      },
...
      "V2ComparisonRequest": {
        "additionalProperties": false,
        "properties": {
          "country_codes": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 2,
            "title": "Country Codes",
            "type": "array"
          },
          "feasibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TfcAssessmentSelectionRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "opportunity_filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityFilterSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "preference_preset_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference Preset Id"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "required": [
          "country_codes"
        ],
        "title": "V2ComparisonRequest",
        "type": "object"
      },
...
      "V2WeightSelection": {
        "additionalProperties": false,
        "properties": {
          "feasibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TfcAssessmentSelectionRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "opportunity_filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityFilterSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "preference_preset_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference Preset Id"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "title": "V2WeightSelection",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Structured API over one immutable schema-current recommendation snapshot.",
    "title": "Konsider API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v3/catalog": {
      "get": {
        "operationId": "catalog_v3_api_v3_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogV3Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Active release unavailable"
          }
        },
        "summary": "Retrieve criteria with uniform display metadata"
      }
    },
    "/api/v3/comparisons": {
      "post": {
        "operationId": "comparisons_v3_api_v3_comparisons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2ComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonV3Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Active release unavailable"
          }
        },
        "summary": "Compare countries using the API v3 display contract"
      }
    },
    "/api/v3/countries/{country_code}/details": {
      "post": {
        "operationId": "country_details_v3_api_v3_countries__country_code__details_post",
        "parameters": [
          {
            "in": "path",
            "name": "country_code",
            "required": true,
            "schema": {
              "title": "Country Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2WeightSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDetailsV3Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Active release unavailable"
          }
        },
        "summary": "Retrieve country evidence using the API v3 display contract"
      }
    },
...
    "/api/v3/opportunity-filters": {
      "get": {
        "operationId": "opportunity_filters_v3_api_v3_opportunity_filters_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityFilterCatalogV3Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Active release unavailable"
          }
        },
        "summary": "Retrieve catalog-grouped filters with uniform display metadata"
      }
    },
    "/api/v3/rankings": {
      "post": {
        "operationId": "rankings_v3_api_v3_rankings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V2RankingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingV3Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Active release unavailable"
          }
        },
        "summary": "Rank countries using the API v3 display contract"
      }
    },
    "/api/v3/tfcs": {
      "get": {
        "operationId": "tfcs_v3_api_v3_tfcs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TfcCatalogV3Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Active release unavailable"
          }
        },
        "summary": "Retrieve feasibility checks with uniform display metadata"
      }
    }
  }
}